feat: extract file manager component
