+++ desktop/src/app/ui.rs
@@ -1,10 +1,7 @@
-use std::collections::HashMap;
 use std::ops::Range;
-use std::path::PathBuf;
 
 use chrono::NaiveDateTime;
 use iced::advanced::text::highlighter::{self, Highlighter};
-use iced::widget::overlay::menu;
 use iced::widget::svg::{Handle, Svg};
 use iced::widget::{
     button, checkbox, column, container, row, scrollable, text, text_editor, text_input,
@@ -19,42 +16,8 @@ use syntect::parsing::SyntaxSet;
 
 use crate::app::diff::DiffView;
 use crate::app::events::Message;
-use crate::app::{command_palette::COMMANDS, EntryType, FileEntry, MulticodeApp};
+use crate::app::{command_palette::COMMANDS, MulticodeApp};
 use crate::modal::Modal;
-
-#[derive(Debug)]
-pub struct ContextMenu {
-    pub path: PathBuf,
-    pub state: std::cell::RefCell<menu::State>,
-    pub hovered: std::cell::RefCell<Option<usize>>,
-}
-
-impl ContextMenu {
-    pub fn new(path: PathBuf) -> Self {
-        Self {
-            path,
-            state: std::cell::RefCell::new(menu::State::new()),
-            hovered: std::cell::RefCell::new(None),
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum ContextMenuItem {
-    Open,
-    Rename,
-    Delete,
-}
-
-impl ToString for ContextMenuItem {
-    fn to_string(&self) -> String {
-        match self {
-            ContextMenuItem::Open => "Открыть".into(),
-            ContextMenuItem::Rename => "Переименовать".into(),
-            ContextMenuItem::Delete => "Удалить".into(),
-        }
-    }
-}
 
 static SYNTAX_SET: Lazy<SyntaxSet> = Lazy::new(SyntaxSet::load_defaults_newlines);
 pub static THEME_SET: Lazy<ThemeSet> = Lazy::new(ThemeSet::load_defaults);
@@ -63,19 +26,6 @@ const OPEN_ICON: &[u8] = include_bytes!("../../assets/open.svg");
 const SAVE_ICON: &[u8] = include_bytes!("../../assets/save.svg");
 const FORMAT_ICON: &[u8] = include_bytes!("../../assets/format.svg");
 const AUTOCOMPLETE_ICON: &[u8] = include_bytes!("../../assets/autocomplete.svg");
-const FILE_ICON: &[u8] = include_bytes!("../../assets/file.svg");
-const FILE_TEXT_ICON: &[u8] = include_bytes!("../../assets/file-text.svg");
-const FILE_RUST_ICON: &[u8] = include_bytes!("../../assets/file-rust.svg");
-
-static EXT_ICON_MAP: Lazy<HashMap<&'static str, &'static [u8]>> = Lazy::new(|| {
-    let mut m = HashMap::new();
-    m.insert("rs", FILE_RUST_ICON);
-    m.insert("md", FILE_TEXT_ICON);
-    m.insert("txt", FILE_TEXT_ICON);
-    m.insert("json", FILE_TEXT_ICON);
-    m.insert("toml", FILE_TEXT_ICON);
-    m
-});
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct SyntaxSettings {
@@ -505,80 +455,6 @@ impl MulticodeApp {
         }
     }
 
-    pub fn view_entries(&self, entries: &[FileEntry], depth: u16) -> Element<Message> {
-        let mut rows = Vec::new();
-        for entry in entries {
-            let indent = Space::with_width(Length::Fixed((depth * 20) as f32));
-            match &entry.ty {
-                EntryType::File => {
-                    let name = entry
-                        .path
-                        .file_name()
-                        .unwrap()
-                        .to_string_lossy()
-                        .to_string();
-                    let ext = entry
-                        .path
-                        .extension()
-                        .and_then(|e| e.to_str())
-                        .unwrap_or("");
-                    let icon = Svg::new(Handle::from_memory(
-                        EXT_ICON_MAP.get(ext).copied().unwrap_or(FILE_ICON),
-                    ))
-                    .width(Length::Fixed(16.0))
-                    .height(Length::Fixed(16.0));
-                    let content = row![icon, text(name)]
-                        .spacing(5)
-                        .align_items(Alignment::Center);
-                    let row = row![
-                        indent,
-                        MouseArea::new(
-                            button(content).on_press(Message::SelectFile(entry.path.clone())),
-                        )
-                        .on_right_press(Message::ShowContextMenu(entry.path.clone())),
-                    ]
-                    .spacing(5)
-                    .align_items(Alignment::Center)
-                    .height(Length::Fixed(20.0))
-                    .into();
-                    rows.push(row);
-                }
-                EntryType::Dir => {
-                    let name = entry
-                        .path
-                        .file_name()
-                        .unwrap()
-                        .to_string_lossy()
-                        .to_string();
-                    let expanded = self.expanded_dirs.contains(&entry.path);
-                    let icon = if expanded { "▼" } else { "▶" };
-                    let content = row![text(icon), text(name)]
-                        .spacing(5)
-                        .align_items(Alignment::Center);
-                    let header = row![
-                        indent,
-                        MouseArea::new(
-                            button(content).on_press(Message::ToggleDir(entry.path.clone())),
-                        )
-                        .on_right_press(Message::ShowContextMenu(entry.path.clone())),
-                    ]
-                    .spacing(5)
-                    .align_items(Alignment::Center)
-                    .height(Length::Fixed(20.0))
-                    .into();
-                    rows.push(header);
-                    if expanded {
-                        rows.push(self.view_entries(&entry.children, depth + 1));
-                    }
-                }
-            }
-        }
-        column(rows).into()
-    }
-
-    pub fn file_tree(&self) -> Element<Message> {
-        scrollable(self.view_entries(&self.files, 0)).into()
-    }
 
     pub fn status_bar_component(&self) -> Element<Message> {
         if !self.settings.show_status_bar {
@@ -694,9 +570,10 @@ impl MulticodeApp {
 #[cfg(test)]
 mod tests {
     use super::*;
-
     use super::super::{CreateTarget, MulticodeApp, Screen, UserSettings};
+    use crate::components::file_manager::ContextMenu;
     use std::collections::HashSet;
+    use std::path::PathBuf;
     use tokio::sync::broadcast;
 
     #[test]
