+++ desktop/src/components/mod.rs
@@ -1,0 +1,3 @@
+pub mod file_manager;
+
+pub use file_manager::{file_tree, view_entries, ContextMenu, ContextMenuItem};
