+++ desktop/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod app;
+pub mod components;
+pub mod modal;
